customer_segmentation/requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
umap-learn